# swagger-full.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "图书馆座位预约系统 API",
        "version": "1.0.0",
...
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "开发环境"
        },
        {
            "url": "http://localhost:3001",
            "description": "测试环境"
        },
        {
            "url": "http://localhost:3002",
            "description": "UAT 环境"
        },
        {
            "url": "http://localhost:3003",
            "description": "生产环境"
        }
    ],
...
    "paths": {
...
        "/api/user/profile": {
            "get": {
                "tags": ["👤 用户管理 (User)"],
                "summary": "获取个人信息",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未授权"
                    },
                    "404": {
                        "description": "用户不存在",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ],
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 3001
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["👤 用户管理 (User)"],
                "summary": "更新个人信息",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "nickName": {
                                        "type": "string"
                                    },
                                    "avatarUrl": {
                                        "type": "string",
                                        "format": "uri"
                                    },
                                    "studentId": {
                                        "type": "string"
                                    },
                                    "realName": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "更新成功"
                    },
                    "400": {
                        "description": "参数错误"
                    }
                }
            }
        },
        "/api/user/favorite/{seatId}": {
            "post": {
                "tags": ["👤 用户管理 (User)"],
                "summary": "收藏/取消收藏座位",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "seatId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "座位 ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功"
                    },
                    "401": {
                        "description": "未授权"
                    },
                    "404": {
                        "description": "座位不存在",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ],
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 4001
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/credit": {
            "get": {
                "tags": ["👤 用户管理 (User)"],
                "summary": "获取信用积分",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "creditScore": {
                                                    "type": "integer",
                                                    "example": 100
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未授权"
                    }
                }
            }
        },
        "/api/user/credit/records": {
            "get": {
                "tags": ["👤 用户管理 (User)"],
                "summary": "获取信用记录",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        },
                        "description": "页码"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        },
                        "description": "每页数量"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "list": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "id": {
                                                                "type": "string"
                                                            },
                                                            "userId": {
                                                                "type": "string"
                                                            },
                                                            "userName": {
                                                                "type": "string"
                                                            },
                                                            "userAvatar": {
                                                                "type": "string"
                                                            },
                                                            "type": {
                                                                "type": "integer"
                                                            },
                                                            "points": {
                                                                "type": "integer"
                                                            },
                                                            "reason": {
                                                                "type": "string"
                                                            },
                                                            "date": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "updatedBy": {
                                                                "type": "string"
                                                            },
                                                            "updatedByName": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未授权"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Token 认证，格式：Bearer <token>"
            }
        },
        "schemas": {
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "type": "object",
                        "properties": {
                            "code": {
                                "type": "integer",
                                "example": 1001,
                                "description": "数字错误码"
                            },
                            "message": {
                                "type": "string",
                                "example": "错误描述"
                            }
                        }
                    }
                },
                "required": ["success", "error"]
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "objectId",
                        "description": "用户 ID"
                    },
                    "username": {
                        "type": "string",
                        "description": "用户名"
                    },
                    "name": {
                        "type": "string",
                        "nullable": true,
                        "description": "用户姓名"
                    },
                    "avatar": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true,
                        "description": "头像 URL"
                    },
                    "email": {
                        "type": "string",
                        "nullable": true,
                        "description": "电子邮件"
                    },
                    "phone": {
                        "type": "string",
                        "nullable": true,
                        "description": "手机号"
                    },
                    "studentId": {
                        "type": "string",
                        "nullable": true,
                        "description": "学号"
                    },
                    "role": {
                        "type": "integer",
                        "enum": [0, 1],
                        "description": "用户角色"
                    },
                    "isSuperAdmin": {
                        "type": "boolean",
                        "description": "是否超级管理员"
                    },
                    "creditScore": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 100,
                        "description": "信用积分"
                    },
                    "blacklisted": {
                        "type": "boolean",
                        "description": "是否黑名单用户"
                    },
                    "blacklistReason": {
                        "type": "string",
                        "nullable": true,
                        "description": "黑名单原因"
                    },
                    "favorites": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "description": "收藏座位 ID 列表"
                    },
                    "activityRegistrations": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "activity": {
                                    "type": "string",
                                    "description": "活动 ID"
                                },
                                "registeredAt": {
                                    "type": "string",
                                    "format": "date-time",
                                    "description": "报名时间"
                                }
                            }
                        },
                        "description": "活动报名记录"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "创建时间"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "更新时间"
                    }
                }
            },
...
        }
    }
}